Maybe you can add commas between the "like a dog" cause a bit confusing?
Hi, this is karena :) Idk if im sending this correctly so im just trying out..... pls tell me if i did this wrongly :/ thanks... Also, I think you wrote a really great story! :D It's entertaining to read and I like the unusual setting as well as the plot twist in the middle!